Refactor: Move collision detection data structures to collision.rs
